[TM-1105] Finished refactoring the Nursery PD page.
